Production runs + output QC: Start/Finish flow, per-pack output, decimal units
Adds the desktop Production-runs tab (Start + Finish), the per-pack
Finish dialog that captures dimensions / weight / stack factor per
physical package the operator produced, and the separate Output-QC
sign-off step that gates manufactured lots from reaching the
warehouse until a production-QC operator passes / fails them.

Run flow (Start / Finish)
- MO gains actual_start, actual_finish, quantity_produced,
  produced_lot_id (migration 20260622180000). Schema run_changeset
  guards finish ≥ start + quantity_produced ≥ 0.
- start_mo_production: scheduled + preflight-clear → in_progress,
  stamps actual_start = now(). Idempotent.
- finish_mo_production: in_progress → completed, creates one
  output stock_lot per pack at the production-feed cell with
  status=received + source_kind=manufacturing_order, also writes
  per-step actual_start / actual_finish from the divider slider
  the FE produces. Captures manufactured_at = actual_finish and
  expiry_at = manufactured_at + item.shelf_life_months (read from
  finished_product_spec / raw_material_compliance; semi-finished
  and packaging keep expiry nil).
- New routes: GET /api/production/runs, POST
  /manufacturing-orders/:id/start-production, POST
  /manufacturing-orders/:id/finish-production. production_run_entry
  payload for the queue, run fields added to manufacturing_order
  payload.

Per-pack output
- One physical pack = one output lot, mirroring the PO-receive
  shape. Sum of pack qtys must equal quantity_produced. Each pack
  carries qty + L/W/H mm + weight kg + stack factor; the BE sets
  units_per_package = qty so volume math always resolves to 1
  package per lot. Hides units_per_package from the UI to avoid
  the old 25 kg → 625 kg multiplier surprise.
- units_per_package migrated from integer to numeric(10,3)
  (20260622220000) so continuous UoM items can record fractional
  pack contents (4.4 kg / bag) without truncation. Downstream call
  sites updated: stock.compute_lot_footprint accepts %Decimal{},
  PO receive validator + Goods-In Inspection pack validator switch
  to parse_positive_decimal, manual receive form drops its ceiling
  hack, packaging_average uses decimal_median.

Output QC
- New production.qc_output capability + matrix row (single
  Update-column workflow checkbox).
- list_pending_output_qc + sign_off_output_qc context fns; the
  sign-off wraps Stock.Lifecycle.record_event with qc_passed /
  qc_failed so the lot's lifecycle ledger remains the only place
  status flips. Refuses non-manufacturing_order lots so the
  production-QC role doesn't bleed into incoming-PO inspections.
- New routes GET /api/production/output-qc + POST .../:lot_uuid.
- Desktop /production/output-qc subnav tab + overview tile. Cards
  show item, qty + UoM, source MO, package spec strip, cell
  breadcrumb. Pass = one click → status flips to available; Fail =
  two-step with required reason → status flips to qc_failed.

Run detail page (shared with MO detail)
- Uses the canonical MOPartsTable / MOOperationsTable / MOCostSummary
  / MOChainRoadmap so the layout matches /production/manufacturing-
  orders/[uuid]. Adds the stage banner (preflight-pending / ready /
  running / completed), the run controls card with Start/Finish, a
  4-up context grid (materials arrived + picker name, production-
  feed cell, planned bounds, run times), and the production-feed
  cell panel with FloorPlanMini + breadcrumb (warehouse → floor →
  location → cell).
- Finish dialog: sticky header + footer, scrollable body. Editable
  start + finish times, produced qty, output packages list (1 by
  default with + Add pack / Remove), and a horizontal operation-
  time divider with grip handles + ±2% keyboard nudge that slices
  the run span across MO steps. Submitted shape includes packs +
  operation_times so the BE stamps per-step actual_start / finish
  in the same transaction that closes the MO.

Stock lot detail polish
- Source reference is now a deep link: "Open MO" → /production/
  manufacturing-orders/<uuid>, "Open PO" → /procurement/purchase-
  orders/<uuid>, with a truncated UUID for visual identification.
  Non-routable kinds (manual / adjustment / opening_balance / return)
  keep the plain mono text.
- MOPartsTable Storage column now shows the rack code (SL00023)
  above the cell level so the picker can find the physical rack at
  a glance — "Level 0" alone never said WHERE.

Permission matrix
- Production section now has its workflow capabilities as
  dedicated single-checkbox rows: Prepare MO, Approve MO, Release
  MO, Pick from warehouse, Confirm receipt at production line, Run
  production (Start / Finish), Quality-check finished output. The
  matrix grid renders them with "—" in the unused columns. Closes
  the older gap where mo_prepare / mo_approve / mo_execute existed
  but weren't surfaced in the role-editor UI.

+++ backend/lib/backend/stock/lot.ex
@@ -71,7 +71,11 @@ defmodule Backend.Stock.Lot do
     field :package_width_mm, :integer
     field :package_height_mm, :integer
     field :package_weight_kg, :decimal
-    field :units_per_package, :integer, default: 1
+    # Decimal (numeric(10,3)) so continuous-UoM items can store
+    # fractional values — e.g. a 4.4 kg produced bag has
+    # units_per_package=4.4. Integer-only values (24 cans, 100
+    # capsules) cast losslessly into the same column.
+    field :units_per_package, :decimal, default: 1
     field :stack_factor, :integer, default: 1
 
     belongs_to :company, Company
